Only redirect if CiviCRM is not already installed
If you install via CLI you will get redirected to setup page first time you go to CiviCRM in admin menu - fine if you did not do a CLI install `cv core:install` But if you did you get a you are not allowed to access this page message.

Need to add a check to see if `CIVICRM_INSTALLED` before doing the redirect.  This will also cover deactivating and reactivating for testing.