HookTest - Fix execution on PHP 8
Before
------

Test fails on PHP 8

```
CiviWP\HookTest::testFoo
Undefined array key "hook_was_called"

/home/jenkins/bknix-dfl/build/wp-265-47vsb/web/wp-content/plugins/civicrm/tests/phpunit/CiviWP/HookTest.php:22
```

After
-----

Test passes

Comments
--------

This is a superficial change - the point of the assertion is to validate the
starting circumstance (asserting that the hook has not yet run). This is just
a cleaner assertion that the hook has not yet run.